+++ src/pools/worker.ts
@@ -2,6 +2,7 @@ import type { EventEmitter } from 'node:events'
 import type { MessageChannel, WorkerOptions } from 'node:worker_threads'
 
 import type { CircularArray } from '../circular-array.js'
+import type { Deque } from '../deque.js'
 import type { Task } from '../utility-types.js'
 
 /**
@@ -300,6 +301,10 @@ export interface IWorkerNode<Worker extends IWorker, Data = unknown>
    */
   readonly messageChannel?: MessageChannel
   /**
+   * Tasks queue.
+   */
+  readonly tasksQueue: Deque<Task<Data>>
+  /**
    * Tasks queue back pressure size.
    * This is the number of tasks that can be enqueued before the worker node has back pressure.
    */
@@ -336,6 +341,12 @@ export interface IWorkerNode<Worker extends IWorker, Data = unknown>
    * @returns The popped task.
    */
   readonly popTask: () => Task<Data> | undefined
+  /**
+   * Deletes a task from the tasks queue.
+   *
+   * @returns `true` if the task was deleted, `false` otherwise.
+   */
+  readonly deleteTask: (task: Task<Data>) => boolean
   /**
    * Clears tasks queue.
    */
@@ -398,4 +409,5 @@ export interface IWorkerNode<Worker extends IWorker, Data = unknown>
 export interface WorkerNodeEventDetail {
   workerId?: number
   workerNodeKey?: number
+  taskId?: string
 }