+++ src/utility-types.ts
@@ -88,6 +88,10 @@ export interface Task<Data = unknown> {
    */
   readonly timestamp?: number
   /**
+   * Whether the task is abortable or not.
+   */
+  readonly abortable?: boolean
+  /**
    * Task UUID.
    */
   readonly taskId?: string
@@ -114,6 +118,11 @@ export interface MessageValue<Data = unknown, ErrorData = unknown>
    * Worker error.
    */
   readonly workerError?: WorkerError<ErrorData>
+  /**
+   * Task operation:
+   * - `'abort'` - Abort a task.
+   */
+  readonly taskOperation?: 'abort'
   /**
    * Task performance.
    */
@@ -182,6 +191,10 @@ export interface PromiseResponseWrapper<Response = unknown> {
    * The asynchronous resource used to track the task execution.
    */
   readonly asyncResource?: AsyncResource
+  /**
+   * The task abort signal.
+   */
+  readonly abortSignal?: AbortSignal
 }
 
 export type Writable<T> = { -readonly [P in keyof T]: T[P] }