+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix publishing on JSR, take 3.
+
 ## [3.1.25] - 2024-03-27
 
 ### Fixed