+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure the minimum number of workers on a started pool is guaranteed.
+
 ## [3.1.27] - 2024-03-27
 
 ### Fixed